% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterGotchaPhasedMitoCalls.R
\name{FilterGotchaPhasedMitoCalls}
\alias{FilterGotchaPhasedMitoCalls}
\title{Select mitochondrial mutations in phase with the got-cha targeted mutation}
\usage{
FilterGotchaPhasedMitoCalls(mitomut_table)
}
\arguments{
\item{mitomut_table}{Dataframe with all mitomutations identified}
}
\value{
Dataframe with mutations which can be used to enhance the Got-cha genotyping efficiency.
}
\description{
Select mitochondrial mutations in phase with the got-cha targeted mutation
}
