% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyGotchaPhasedMitoCalls.R
\name{IdentifyGotchaPhasedMitoCalls}
\alias{IdentifyGotchaPhasedMitoCalls}
\title{Identify mitochondrial mutations in phase with the got-cha targeted mutation}
\usage{
IdentifyGotchaPhasedMitoCalls(
  mgatk,
  genotyping_table,
  gene,
  min_cellcov_calculate_af = 10,
  pattern = "-1",
  frac_thres = 1,
  numcellsupportingvar = 5,
  strand_correlation = 0.65,
  log10_vmr = -2,
  low_coverage_threshold = 10,
  stabilize_variance = TRUE
)
}
\arguments{
\item{mgatk}{RangedSummarizedExperiment object which can be generated following the pipeline in \link[mgatk]{https://github.com/caleblareau/mgatk}.}

\item{genotyping_table}{Got-cha genotyping table.}

\item{gene}{Name of the targeted gene of interest. It has to match the column name within the got-cha genotyping table.}

\item{min_cellcov_calculate_af}{Minimum cell coverage at a given site to generate an heteroplasmic value.}

\item{pattern}{Suffix added to all the cell/sample names.}

\item{frac_thres}{Minimum percentage of cells required in at least two out of the three categories of cells containing 0\% (1), 100\% (2) or intermediate heteroplasmic values (3).}
}
\value{
Dataframe with mutations which can be used to enhance the Got-cha genotyping efficiency.
}
\description{
Identify mitochondrial mutations in phase with the got-cha targeted mutation
}
\examples{
load("data/Input.RData", verbose = T)
IdentifyGotchaPhasedMitoCalls(mgatk=mgatk, genotyping_table = genotyping_table, gene = "JAK2")
}
