% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingDataCum.R
\name{MissingDataCum}
\alias{MissingDataCum}
\title{Optimize the number of variants to use in the classifier given their percentage of missing data}
\usage{
MissingDataCum(nsites, mitomutations, pattern = "-1")
}
\arguments{
\item{nsites}{Number of sites to consider. Between 1 and the total number of sites identified as in phase.}

\item{mitomutations}{Table containing mitomutations and annotations about their informativeness.}
}
\value{
Dataframe with mutations which can be used to enhance the Got-cha genotyping efficiency.
}
\description{
Optimize the number of variants to use in the classifier given their percentage of missing data
}
\examples{

}
