% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MitoClassifier.R
\name{MitoClassifier}
\alias{MitoClassifier}
\title{Build and apply a random forest classifier using in phase mitocondrial mutations}
\usage{
MitoClassifier(
  mitomutations,
  training_prop = 0.75,
  number_trees = 1000L,
  genotypequal_threshold = 0.5,
  outdir = "data/results/"
)
}
\arguments{
\item{mitomutations}{Table containing mitomutations and annotations about their informativeness.}

\item{training_prop}{Proportion of the data to be used for training the classifier.}
}
\value{
Dataframe with Original GoT-ChA genotype together with the classifier predictions.
}
\description{
Build and apply a random forest classifier using in phase mitocondrial mutations
}
